Feature backend springsecurity

+++ src/main/java/com/example/swappy/service/UserService.java
@@ -1,8 +1,11 @@
 package com.example.swappy.service;
 
+import com.example.swappy.exception.user.NoSuchUserExistsException;
+import com.example.swappy.exception.user.UserAlreadyExistsException;
 import com.example.swappy.model.User;
 import com.example.swappy.repository.UserRepository;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 
 @Service
@@ -19,15 +22,28 @@ public class UserService {
     }
 
     public User getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+
+        User user = userRepository.findById(id).orElse(null);
+
+        if (user == null) {
+            throw new NoSuchUserExistsException("User with id " + id + " not found!");
+        } else return user;
     }
 
     public User saveUser(User user) {
-        return userRepository.save(user);
+        if (userRepository.findByEmail(user.getEmail()) != null) {
+            throw new UserAlreadyExistsException("User already exists!");
+        } else {
+            return userRepository.save(user);
+        }
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.findById(id).orElse(null) == null) {
+            throw new NoSuchUserExistsException("User with id " + id + " not found!");
+        } else {
+            userRepository.deleteById(id);
+        }
     }
 }
